+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Use per request configuration when cache is disabled [PR #289](https://github.com/3scale/apicast/pull/289)
+
 ### Added
 
 ### Removed